+++ tests/smoke_update.rs
@@ -1,14 +1,14 @@
+use std::fs::{self, File};
+
+use flate2::{write::GzEncoder, Compression};
 use kingfisher::{cli::global::GlobalArgs, update::check_for_update};
+use tar::Builder;
+use tempfile::tempdir;
 use tokio;
 use wiremock::{
     matchers::{method, path},
     Mock, MockServer, ResponseTemplate,
 };
-
-use flate2::{write::GzEncoder, Compression};
-use tar::Builder;
-use tempfile::tempdir;
-use std::fs::{self, File};
 
 #[tokio::test]
 async fn no_update_when_flag_set() {